requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.87.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.2
pydantic==1.10.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
yarl==1.8.1
